split importer and move to whyis package
